Add MergedPRsMetric
Uses the same underlying graphql query that the TimeToMergeMetric uses.

Remove the RepositoryReport#mergedPullRequests property -- it does not
capture all merged PRs in the interval because it constrains to only PRs
that were created in the interval. It's possible for a PR to be created before
the start of the interval but merged during the interval -- PRs matching this
criteria were not being included.